backend/requirements.txt
Flask
Flask-Migrate
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn
flask_cors
Pillow
